Added support for "query" action to fetch welcome message.
Send an audio packet of silence to ask for welcome message.
Disconnect when finished, so we don't show up in the list.
Added some new message types.
Changed ${len} to {$len} in a few places.
Unset $server reference when it is finished with.
Added a $done flag to avoid timeout wait when not needed.